handle event calls plus pastabot
Used tasks to call the handlers from the Program.py.
This is very basic and assumes the handle implemented the on_message
method.
this is temporary and will be changed to something better.
You need to create a class that implements it the same way (look at
pastabot or hofer handlers)

Also pastabot v0.1 is out.